Membuat auth dan perlu dirapihkan lagi

+++ routes/web.php
@@ -34,42 +34,64 @@ Route::get('/populer', function () {
 //---------------------backend----------------------
 Route::get('/dashboard', function () {
     return view('backend.dashboard');
-});
+})->middleware('auth');
 
-Route::get('/buku', function () {
+/*Route::get('/buku', function () {
     return view('backend.');
 });
-
+*/
 //Route::resource('/buku', BukuController::class);
 //Route::resource('/kategori', KategoriController::class);
 Route::resource('/meminjam', MeminjamController::class);
 //Route::resource('/penerbit', PenerbitController::class);
-Route::resource('/pengarangbuku', PengarangbukuController::class);
+//Route::resource('/pengarangbuku', PengarangbukuController::class);
 // Route::resource('/pengarang',PengarangController::class);
-Route::resource('/salinanbuku', SalinanbukuController::class);
+//Route::resource('/salinanbuku', SalinanbukuController::class);
 
 //-----------------table buku------------------
-Route::get('/buku', [BukuController::class, 'index']);
-Route::post('/buku', [BukuController::class, 'store'])->name('simpandata0');
-Route::put('/buku/{idbuku}', [BukuController::class, 'update'])->name('ubahdata0');
-Route::delete('/buku/{idbuku}', [BukuController::class, 'destroy'])->name('hapusdata0');
+Route::get('/buku', [BukuController::class, 'index'])->middleware('auth');
+Route::post('/buku', [BukuController::class, 'store'])->name('simpandata0')->middleware('auth');
+Route::put('/buku/{idbuku}', [BukuController::class, 'update'])->name('ubahdata0')->middleware('auth');
+Route::delete('/buku/{idbuku}', [BukuController::class, 'destroy'])->name('hapusdata0')->middleware('auth');
 
 //-----------------table kategori------------------
-Route::get('/kategori', [KategoriController::class, 'index']);
-Route::post('/kategori', [KategoriController::class, 'store'])->name('simpandata1');
-Route::put('/kategori/{idkategori}', [KategoriController::class, 'update'])->name('ubahdata1');
-Route::delete('/kategori/{idkategori}', [KategoriController::class, 'destroy'])->name('hapusdata1');
+Route::get('/kategori', [KategoriController::class, 'index'])->middleware('auth');
+Route::post('/kategori', [KategoriController::class, 'store'])->name('simpandata1')->middleware('auth');
+Route::put('/kategori/{idkategori}', [KategoriController::class, 'update'])->name('ubahdata1')->middleware('auth');
+Route::delete('/kategori/{idkategori}', [KategoriController::class, 'destroy'])->name('hapusdata1')->middleware('auth');
 
 
 //-----------------table pengarang------------------
-Route::get('/pengarang', [PengarangController::class, 'index']);
-Route::post('/pengarang', [PengarangController::class, 'store'])->name('simpandata');
-Route::put('/pengarang/{idpengarang}', [PengarangController::class, 'update'])->name('ubahdata');
-Route::delete('/pengarang/{idpengarang}', [PengarangController::class, 'destroy'])->name('hapusdata');
+Route::get('/pengarang', [PengarangController::class, 'index'])->middleware('auth');
+Route::post('/pengarang', [PengarangController::class, 'store'])->name('simpandata')->middleware('auth');
+Route::put('/pengarang/{idpengarang}', [PengarangController::class, 'update'])->name('ubahdata')->middleware('auth');
+Route::delete('/pengarang/{idpengarang}', [PengarangController::class, 'destroy'])->name('hapusdata')->middleware('auth');
 
 
 //-----------------table penerbit------------------
-Route::get('/penerbit', [PenerbitController::class, 'index']);
-Route::post('/penerbit', [PenerbitController::class, 'store'])->name('simpandata2');
-Route::put('/penerbit/{idpenerbit}', [PenerbitController::class, 'update'])->name('ubahdata2');
-Route::delete('/penerbit/{idpenerbit}', [PenerbitController::class, 'destroy'])->name('hapusdata2');
+Route::get('/penerbit', [PenerbitController::class, 'index'])->middleware('auth');
+Route::post('/penerbit', [PenerbitController::class, 'store'])->name('simpandata2')->middleware('auth');
+Route::put('/penerbit/{idpenerbit}', [PenerbitController::class, 'update'])->name('ubahdata2')->middleware('auth');
+Route::delete('/penerbit/{idpenerbit}', [PenerbitController::class, 'destroy'])->name('hapusdata2')->middleware('auth');
+
+//-----------------table kelola akun------------------
+Route::get('/users', [PenerbitController::class, 'index'])->middleware('auth');
+Route::post('/users', [PenerbitController::class, 'store'])->name('simpandata2')->middleware('auth');
+Route::put('/users/{idusers}', [PenerbitController::class, 'update'])->name('ubahdata2')->middleware('auth');
+Route::delete('/users/{idusers}', [PenerbitController::class, 'destroy'])->name('hapusdata2')->middleware('auth');
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+
+Route::get('/afterregister', function () {
+    return view('backend.afterRegister');
+});
+
+Route::get('/users', function () {
+    return view('backend.users');
+});
+
+Route::get('/profile', function () {
+    return view('backend.profile');
+});
